Ignore missing COUT and CERR variables

+++ src/main/groovy/nextflow/executor/BashWrapperBuilder.groovy
@@ -106,8 +106,9 @@ class BashWrapperBuilder {
         on_exit() {
           exit_status=${ret:=$?}
           printf $exit_status > __EXIT_FILE__
-          rm -f "$COUT" || true
-          rm -f "$CERR" || true
+          set +u
+          [[ "$COUT" ]] && rm -f "$COUT" || true
+          [[ "$CERR" ]] && rm -f "$CERR" || true
           exit $exit_status
         }
