+++ src/test/groovy/nextflow/executor/BashWrapperBuilderTest.groovy
@@ -164,8 +164,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -270,8 +271,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -472,8 +474,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -595,8 +598,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -801,8 +805,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -908,8 +913,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -1018,8 +1024,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -1124,8 +1131,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -1234,8 +1242,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -1373,8 +1382,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
 
@@ -1477,8 +1487,9 @@ class BashWrapperBuilderTest extends Specification {
                     on_exit() {
                       exit_status=${ret:=$?}
                       printf $exit_status > /my/exit/file\\'s
-                      rm -f "$COUT" || true
-                      rm -f "$CERR" || true
+                      set +u
+                      [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                      [[ "\$CERR" ]] && rm -f "\$CERR" || true
                       exit $exit_status
                     }
 
@@ -1528,8 +1539,9 @@ class BashWrapperBuilderTest extends Specification {
                     on_exit() {
                       exit_status=${ret:=$?}
                       printf $exit_status > /my/exit/xxx
-                      rm -f "$COUT" || true
-                      rm -f "$CERR" || true
+                      set +u
+                      [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                      [[ "\$CERR" ]] && rm -f "\$CERR" || true
                       exit $exit_status
                     }
 
@@ -1684,8 +1696,9 @@ class BashWrapperBuilderTest extends Specification {
                 on_exit() {
                   exit_status=\${ret:=\$?}
                   printf \$exit_status > ${folder}/.exitcode
-                  rm -f "\$COUT" || true
-                  rm -f "\$CERR" || true
+                  set +u
+                  [[ "\$COUT" ]] && rm -f "\$COUT" || true
+                  [[ "\$CERR" ]] && rm -f "\$CERR" || true
                   exit \$exit_status
                 }
